% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_bh_binary_variables}
\alias{add_bh_binary_variables}
\title{add binary variable to the data for bank holidays}
\usage{
add_bh_binary_variables(data, holidays)
}
\arguments{
\item{data}{table with a "date" field in it}

\item{holidays}{date; vector of bank holidays in the period}
}
\description{
adds binary variables for days around bank holidays that aren't
  easter or xmas. Bank holidays that occur on Tuesday, Wednesday or Thursday
  aren't picked out as it is assumed that the variability in registrations
  occurs within the same week that the bank holiday occurred
}
