% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_day_weighting}
\alias{add_day_weighting}
\title{add a weighting to each day for the months it can influence}
\usage{
add_day_weighting(data)
}
\arguments{
\item{data}{tibble containing a "date" field containing every date in the
time period of interest}
}
\description{
adds a weighting score between 0 and 1 to indicate how much of a
  weight the month that day falls into should influence the model
}
