% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_easter_binary_variables}
\alias{add_easter_binary_variables}
\title{add binary variable to the table for Easter}
\usage{
add_easter_binary_variables(data, easter_fridays)
}
\arguments{
\item{data}{table that includes a "date" field}

\item{easter_fridays}{dates of easter fridays within the dataset}
}
\description{
adds binary variables for days around easter holidays
}
