% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_prediction_intervals}
\alias{add_prediction_intervals}
\title{add prediction intervals to the expected value}
\usage{
add_prediction_intervals(data, dispersion_parameter)
}
\arguments{
\item{data}{data frame containing data}

\item{dispersion_parameter}{dispersion parameter from model}

\item{modelled_field}{unquoted field name of field in data containing the
expected values}
}
\description{
add lower and upper prediction interval fields to data frame
}
