% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_prediction_dates}
\alias{build_prediction_dates}
\title{build the prediction table for the model}
\usage{
build_prediction_dates(
  areacode = c("E02000001", "E02000002", "E02000003"),
  deprivation_quintile = 1:5,
  sex = 0:1,
  age_group = c("0-24", "25-49", "50-64", "65-74", "75-84", "85+"),
  from_date,
  to_date,
  holidays,
  denominators
)
}
\arguments{
\item{areacode}{character vector of area codes to generate data frame for}

\item{deprivation_quintile}{numeric; deprivation quintile number (1 is least
deprived quintile and 5 is the most deprived quintile)}

\item{sex}{0 (male) or 1 (female)}

\item{age_group}{string; describing the age groups used for modelling}

\item{from_date}{date; a single date to make predications from}

\item{to_date}{date; a single date to make predictions to}

\item{holidays}{date; vector containing public holidays}

\item{denominators}{tibble containing weekly denominator data for each
areacode, sex, age group, deprivation quintile}
}
\description{
build a data frame for prediction dates with same structure as
  modelled data that model can be used to predict deaths on
}
