% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_easter_fridays}
\alias{calc_easter_fridays}
\title{calculate Easter Fridays from an input of dates}
\usage{
calc_easter_fridays(holidays, include_christmas_friday = TRUE)
}
\arguments{
\item{holidays}{date; vector of holiday dates}

\item{include_christmas_friday}{logical; whether to treat a Christmas on a
Friday in the same was as an Easter Friday (because the following Monday is
also a bank holiday)}
}
\description{
returns a vector of easter fridays when given a vector of
  holiday dates
}
