% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_average_bank_holidays_per_year}
\alias{check_average_bank_holidays_per_year}
\title{Check on holiday input dates}
\usage{
check_average_bank_holidays_per_year(holidays)
}
\arguments{
\item{holidays}{date; vector of bank holidays in period}
}
\description{
Check on holiday input dates
}
