% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_holiday_dates}
\alias{check_holiday_dates}
\title{Check on holiday input dates}
\usage{
check_holiday_dates(holidays)
}
\arguments{
\item{holidays}{date; vector of bank holidays in period}
}
\description{
Check on holiday input dates
}
