% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{check_monthly_populations}
\alias{check_monthly_populations}
\title{Sense check monthly population data}
\usage{
check_monthly_populations(populations)
}
\arguments{
\item{populations}{table of populations containing areacode, age_group, sex,
deprivation_quintile, year and population}
}
\value{
function will error if something is wrong with the data inputs,
  otherwise it will return the sum of the population inputs
}
\description{
checks on the expected number of records, all the names are
  present, each record is unique, there are no NAs and that all population
  values are positive or zero
}
\examples{
pops <- create_dummy_populations() |>
   convert_annual_to_monthly_populations()
pop_checks <- check_monthly_populations(pops)
}
