% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{check_raw_populations}
\alias{check_raw_populations}
\title{Sense check raw population data}
\usage{
check_raw_populations(populations)
}
\arguments{
\item{populations}{table of populations containing areacode, age_group, sex,
deprivation_quintile, year and population}
}
\value{
function will error if something is wrong with the data inputs,
  otherwise it will return the a table of populations by year
}
\description{
checks on the expected number of records, all the names are
  present, each record is unique, there are no NAs and that all population
  values are positive or zero
}
\examples{
pops <- create_dummy_populations()
pop_checks <- check_raw_populations(pops)
}
