% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{christmas_week}
\alias{christmas_week}
\title{calculate the dates in the same week as Christmas}
\usage{
christmas_week(year, day_of_interest)
}
\arguments{
\item{year}{integer; year of interest}

\item{day_of_interest}{integer between 1 and 7 for day of the week (1 is
Sunday)}
}
\description{
function that creates a vector of dates that sit in the same week
  as Christmas, where Christmas day is on a specified day of the week (ie, if
  interested in a Friday Christmas only and the years 2015 to 2020 were
  provided, only 19th to 25th December 2015 would be returned)
}
