% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{consecutive_bank_hols}
\alias{consecutive_bank_hols}
\title{applies variables to weeks where there are two consecutive bank holidays,
that don't include a Monday/Tuesday combination. Where these occur on xmas,
the two following weeks will be identified with additional variables. This
function aims to pick up the Queen's Jubilee weekend}
\usage{
consecutive_bank_hols(data, holidays)
}
\arguments{
\item{data}{tibble containing a date field}

\item{date_field}{unquoted name of the field in data which corresponds to the
dates}
}
\description{
applies variables to weeks where there are two consecutive bank holidays,
that don't include a Monday/Tuesday combination. Where these occur on xmas,
the two following weeks will be identified with additional variables. This
function aims to pick up the Queen's Jubilee weekend
}
