% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{convert_annual_to_monthly_populations}
\alias{convert_annual_to_monthly_populations}
\title{Convert population mid-year estimates into monthly estimates by population
subgroups}
\usage{
convert_annual_to_monthly_populations(denominators, start_year, end_year)
}
\arguments{
\item{denominators}{table that contains areacode, deprivation_quintile, sex,
age_group, period, population}

\item{start_year}{numeric the start year for your output dataset}

\item{end_year}{numeric the end year for your output dataset}
}
\value{
tibble containing all the same input data and also month
}
\description{
Convert population mid-year estimates into monthly estimates by population
subgroups
}
