% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{convert_annual_to_weekly_populations}
\alias{convert_annual_to_weekly_populations}
\title{Convert population mid-year estimates into weekly estimates by population
subgroups, accounting for lower populations in weeks containing bank holidays}
\usage{
convert_annual_to_weekly_populations(
  denominators,
  from_date,
  to_date,
  holidays
)
}
\arguments{
\item{denominators}{table that contains areacode, deprivation_quintile, sex,
age_group, period, population}

\item{from_date}{date; earliest date in the period of concern}

\item{to_date}{date; latest date in the period of concern}

\item{holidays}{date; a vector of dates which correspond to bank holidays in
the period}
}
\value{
tibble containing all the same input data and also week_ending
}
\description{
Convert population mid-year estimates into weekly estimates by population
subgroups, accounting for lower populations in weeks containing bank holidays
}
\examples{
df <- create_dummy_populations() |>
 convert_annual_to_weekly_populations(
   from_date = as.Date("2015-01-03"),
   to_date = as.Date("2019-12-27"),
   holidays = as.Date("2017-12-25"))
}
