% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_date_dependent_variables}
\alias{create_date_dependent_variables}
\title{Create table of weekly date dependent variables}
\usage{
create_date_dependent_variables(from_date, to_date, holidays)
}
\arguments{
\item{from_date}{date; first date in time period}

\item{to_date}{date; final date in the time period}

\item{holidays}{vector of dates that represent national holidays}
}
\description{
Create table of weekly date dependent variables
}
\examples{
vars <- create_date_dependent_variables(
  from_date = as.Date("2020-01-01"),
  to_date = as.Date("2020-12-31"),
  holidays = as.Date(c("2020-01-01", "2020-12-25")))
}
