% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{create_dummy_populations}
\alias{create_dummy_populations}
\title{Create a dummy population dataset for testing}
\usage{
create_dummy_populations(
  areacode = c("E02000001", "E02000002", "E02000003"),
  deprivation_quintile = 1:5,
  sex = 0:1,
  age_group = c("0-24", "25-49", "50-64", "65-74", "75-84", "85+"),
  year = 2014:2024
)
}
\arguments{
\item{areacode}{string; areacodes for the areas included in modelling}

\item{deprivation_quintile}{numeric; deprivation quintile number (1 is least
deprived quintile and 5 is the most deprived quintile)}

\item{sex}{0 (male) or 1 (female)}

\item{age_group}{string; describing the age groups used for modelling}

\item{year}{numeric; the year the population estimate is for}
}
\value{
tibble containing areacode, sex, age_group, deprivation_quintile,
  year, population
}
\description{
Create a dummy population dataset for testing
}
\examples{
df <- create_dummy_populations(
        areacode = c("abc", "def"))
}
