% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_we_bh_denominators}
\alias{remove_we_bh_denominators}
\title{removes denominators from weekend and bank holiday dates. Within the
baseline death period, deaths were very unlikely to be registered on
weekends or bank holidays therefore these days should have consistent
predictions with working days}
\usage{
remove_we_bh_denominators(data, holidays)
}
\arguments{
\item{data}{tibble of data where one field contains the denominator}

\item{holidays}{vector of dates which are bank holidays}
}
\description{
removes denominators from weekend and bank holiday dates. Within the
baseline death period, deaths were very unlikely to be registered on
weekends or bank holidays therefore these days should have consistent
predictions with working days
}
