% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_correct}
\alias{round_correct}
\title{Correct rounding}
\usage{
round_correct(x, n)
}
\arguments{
\item{x}{number to round}

\item{n}{decimal places}
}
\description{
rounding function
}
\details{
from https://stackoverflow.com/questions/12688717/round-up-from-5
}
