% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_to_friday}
\alias{round_to_friday}
\title{converts a date to the previous/following Friday}
\usage{
round_to_friday(dt, direction)
}
\arguments{
\item{dt}{date}

\item{direction}{string; "up" or "down" - which direction to round the date}
}
\description{
converts a date to the previous/following Friday
}
