% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_up_to_saturday}
\alias{round_up_to_saturday}
\title{converts a date to the following Saturday}
\usage{
round_up_to_saturday(dt)
}
\arguments{
\item{dt}{date}
}
\description{
converts a date to the following Saturday
}
