% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rqpois}
\alias{rqpois}
\title{simulate distribution around the modelled value}
\usage{
rqpois(n, mu, dispersion_parameter = NULL, var = NULL)
}
\arguments{
\item{n}{number of simulated values}

\item{mu}{expected value}

\item{dispersion_parameter}{dispersion parameter for the model}

\item{var}{variance}
}
\description{
creates distribution of expected values around the modelled
  value
}
