% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sat_to_mon_xmas}
\alias{sat_to_mon_xmas}
\title{applies variables indicating the days which are in a week where Christmas day
falls on a Saturday, Sunday or Monday. Weeks are Saturday to Friday.}
\usage{
sat_to_mon_xmas(data)
}
\arguments{
\item{data}{tibble containing a "date" field}
}
\description{
applies variables indicating the days which are in a week where Christmas day
falls on a Saturday, Sunday or Monday. Weeks are Saturday to Friday.
}
