% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{weekly_denominators}
\alias{weekly_denominators}
\title{aggregates denominator table containing monthly denominator data to a daily
version, then removes the weekends and bank holidays before aggregating it to
a monthly version}
\usage{
weekly_denominators(denominators, from_date, to_date, holidays)
}
\arguments{
\item{denominators}{tibble; the tibble must contain fields for month,
denominator, age, sex, and deprivation quintile}

\item{from_date}{date; earliest date in the period of interest}

\item{to_date}{date; latest date in the period of interest}

\item{holidays}{date; a vector of dates which correspond to bank holidays in
the period}
}
\description{
aggregates denominator table containing monthly denominator data to a daily
version, then removes the weekends and bank holidays before aggregating it to
a monthly version
}
