% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weekly_holiday_variables}
\alias{weekly_holiday_variables}
\title{creates a weekly table of bank holiday variables for modelling or predicting}
\usage{
weekly_holiday_variables(from_date, to_date, holidays)
}
\arguments{
\item{from_date}{date; first date in time period}

\item{to_date}{date; final date in the time period}

\item{holidays}{vector of dates that represent national holidays}
}
\description{
creates a weekly table of bank holiday variables for modelling or predicting
}
