% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weekly_seasonal_variables}
\alias{weekly_seasonal_variables}
\title{creates table of month variables (ie, seasonal) aggregated to week}
\usage{
weekly_seasonal_variables(from_date, to_date)
}
\arguments{
\item{from_date}{date; first date in time period}

\item{to_date}{date; last date in time period}
}
\description{
creates table of month variables (ie, seasonal) aggregated to week
}
