% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weekly_trend_variable}
\alias{weekly_trend_variable}
\title{creates a weekly table mean number of days in that week since 31st Dec 2016}
\usage{
weekly_trend_variable(from_date, to_date)
}
\arguments{
\item{from_date}{date; first date in time period}

\item{to_date}{date; last date in time period}
}
\description{
creates a weekly table mean number of days in that week since 31st Dec 2016
}
